+++ bpmn_parser/_service_task.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 from lxml.etree import _Element
 
-from bpmn_parser.task import Task, TaskElement
+from bpmn_parser._task import Task, TaskElement
 
 
 @dataclass
