+++ bpmn_parser/exceptions.py
@@ -1,0 +1,8 @@
+class BPMNTagNotFound(Exception):
+    def __init__(self, message: str = 'BPMN tag not found!'):
+        super().__init__(message)
+
+
+class NotBPMNFile(Exception):
+    def __init__(self, message: str = 'The file is not a BPMN!'):
+        super().__init__(message)
